pull velocity update into separate file
