+++ src/incflo_apply_predictor.cpp
@@ -70,8 +70,6 @@ void incflo::ApplyPredictor (bool incremental_projection)
 {
     BL_PROFILE("incflo::ApplyPredictor");
 
-    constexpr Real m_half = Real(0.5);
-
     // We use the new time value for things computed on the "*" state
     Real new_time = m_cur_time + m_dt;
 
@@ -95,13 +93,6 @@ void incflo::ApplyPredictor (bool incremental_projection)
                          w_mac[lev].setBndry(0.0););
         }
     }
-
-    // *************************************************************************************
-    // Allocate space for half-time density
-    // *************************************************************************************
-    Vector<MultiFab> density_nph;
-    for (int lev = 0; lev <= finest_level; ++lev)
-        density_nph.emplace_back(grids[lev], dmap[lev], 1, 1, MFInfo(), Factory(lev));
 
     // Forcing terms
     Vector<MultiFab> vel_forces, tra_forces;
@@ -167,173 +158,26 @@ void incflo::ApplyPredictor (bool incremental_projection)
                             m_cur_time);
 
     // *************************************************************************************
-    // Define local variables for lambda to capture.
-    // *************************************************************************************
-    Real l_dt = m_dt;
-
-    // *************************************************************************************
-    // Update density first
+    // Update density
     // *************************************************************************************
     update_density(StepType::Predictor);
 
     // *************************************************************************************
-    // Update tracer next
+    // Update tracer
     // *************************************************************************************
     update_tracer(StepType::Predictor, tra_forces);
 
     // *************************************************************************************
-    // Define (or if advection_type != "MOL", re-define) the forcing terms, without the viscous terms
-    //    and using the half-time density
+    // Update velocity
     // *************************************************************************************
-    compute_vel_forces(GetVecOfPtrs(vel_forces), get_velocity_old_const(),
-                       GetVecOfConstPtrs(density_nph),
-                       get_tracer_old_const(), get_tracer_new_const());
-
-    // *************************************************************************************
-    // Update the velocity
-    // *************************************************************************************
-    for (int lev = 0; lev <= finest_level; lev++)
-    {
-        auto& ld = *m_leveldata[lev];
-#ifdef _OPENMP
-#pragma omp parallel if (Gpu::notInLaunchRegion())
-#endif
-        for (MFIter mfi(ld.velocity,TilingIfNotGPU()); mfi.isValid(); ++mfi)
-        {
-            Box const& bx = mfi.tilebox();
-            Array4<Real> const& vel = ld.velocity.array(mfi);
-            Array4<Real const> const& dvdt = ld.conv_velocity_o.const_array(mfi);
-            Array4<Real const> const& vel_f = vel_forces[lev].const_array(mfi);
-            Array4<Real const> const& rho_old  = ld.density_o.const_array(mfi);
-            Array4<Real const> const& rho_new  = ld.density.const_array(mfi);
-            Array4<Real const> const& rho_nph  = density_nph[lev].array(mfi);
-
-            if (m_diff_type == DiffusionType::Implicit) {
-
-                if (use_tensor_correction)
-                {
-                    Array4<Real const> const& divtau_o = ld.divtau_o.const_array(mfi);
-                    // Here divtau_o is the difference of tensor and scalar divtau_o!
-                    if (m_advect_momentum) {
-                        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-                        {
-                            AMREX_D_TERM(vel(i,j,k,0) *= rho_old(i,j,k);,
-                                         vel(i,j,k,1) *= rho_old(i,j,k);,
-                                         vel(i,j,k,2) *= rho_old(i,j,k););
-                            AMREX_D_TERM(vel(i,j,k,0) += l_dt*(dvdt(i,j,k,0)+rho_nph(i,j,k)*vel_f(i,j,k,0)+divtau_o(i,j,k,0));,
-                                         vel(i,j,k,1) += l_dt*(dvdt(i,j,k,1)+rho_nph(i,j,k)*vel_f(i,j,k,1)+divtau_o(i,j,k,1));,
-                                         vel(i,j,k,2) += l_dt*(dvdt(i,j,k,2)+rho_nph(i,j,k)*vel_f(i,j,k,2)+divtau_o(i,j,k,2)););
-                            AMREX_D_TERM(vel(i,j,k,0) /= rho_new(i,j,k);,
-                                         vel(i,j,k,1) /= rho_new(i,j,k);,
-                                         vel(i,j,k,2) /= rho_new(i,j,k););
-                        });
-                    } else {
-                        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-                        {
-                            AMREX_D_TERM(vel(i,j,k,0) += l_dt*(dvdt(i,j,k,0)+vel_f(i,j,k,0)+divtau_o(i,j,k,0));,
-                                         vel(i,j,k,1) += l_dt*(dvdt(i,j,k,1)+vel_f(i,j,k,1)+divtau_o(i,j,k,1));,
-                                         vel(i,j,k,2) += l_dt*(dvdt(i,j,k,2)+vel_f(i,j,k,2)+divtau_o(i,j,k,2)););
-                        });
-                    }
-                } else {
-                    if (m_advect_momentum) {
-                        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-                        {
-                            AMREX_D_TERM(vel(i,j,k,0) *= rho_old(i,j,k);,
-                                         vel(i,j,k,1) *= rho_old(i,j,k);,
-                                         vel(i,j,k,2) *= rho_old(i,j,k););
-                            AMREX_D_TERM(vel(i,j,k,0) += l_dt*(dvdt(i,j,k,0)+rho_nph(i,j,k)*vel_f(i,j,k,0));,
-                                         vel(i,j,k,1) += l_dt*(dvdt(i,j,k,1)+rho_nph(i,j,k)*vel_f(i,j,k,1));,
-                                         vel(i,j,k,2) += l_dt*(dvdt(i,j,k,2)+rho_nph(i,j,k)*vel_f(i,j,k,2)););
-                            AMREX_D_TERM(vel(i,j,k,0) /= rho_new(i,j,k);,
-                                         vel(i,j,k,1) /= rho_new(i,j,k);,
-                                         vel(i,j,k,2) /= rho_new(i,j,k););
-                        });
-                    } else {
-                        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-                        {
-                            AMREX_D_TERM(vel(i,j,k,0) += l_dt*(dvdt(i,j,k,0)+vel_f(i,j,k,0));,
-                                         vel(i,j,k,1) += l_dt*(dvdt(i,j,k,1)+vel_f(i,j,k,1));,
-                                         vel(i,j,k,2) += l_dt*(dvdt(i,j,k,2)+vel_f(i,j,k,2)););
-                        });
-                    }
-                }
-            }
-            else if (m_diff_type == DiffusionType::Crank_Nicolson)
-            {
-
-                Array4<Real const> const& divtau_o = ld.divtau_o.const_array(mfi);
-                if (m_advect_momentum) {
-                    amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-                    {
-                        AMREX_D_TERM(vel(i,j,k,0) *= rho_old(i,j,k);,
-                                     vel(i,j,k,1) *= rho_old(i,j,k);,
-                                     vel(i,j,k,2) *= rho_old(i,j,k););
-                        AMREX_D_TERM(vel(i,j,k,0) += l_dt*(dvdt(i,j,k,0)+rho_nph(i,j,k)*vel_f(i,j,k,0)+m_half*divtau_o(i,j,k,0));,
-                                     vel(i,j,k,1) += l_dt*(dvdt(i,j,k,1)+rho_nph(i,j,k)*vel_f(i,j,k,1)+m_half*divtau_o(i,j,k,1));,
-                                     vel(i,j,k,2) += l_dt*(dvdt(i,j,k,2)+rho_nph(i,j,k)*vel_f(i,j,k,2)+m_half*divtau_o(i,j,k,2)););
-                        AMREX_D_TERM(vel(i,j,k,0) /= rho_new(i,j,k);,
-                                     vel(i,j,k,1) /= rho_new(i,j,k);,
-                                     vel(i,j,k,2) /= rho_new(i,j,k););
-                    });
-                } else {
-                    amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-                    {
-                        AMREX_D_TERM(vel(i,j,k,0) += l_dt*(dvdt(i,j,k,0)+vel_f(i,j,k,0)+m_half*divtau_o(i,j,k,0));,
-                                     vel(i,j,k,1) += l_dt*(dvdt(i,j,k,1)+vel_f(i,j,k,1)+m_half*divtau_o(i,j,k,1));,
-                                     vel(i,j,k,2) += l_dt*(dvdt(i,j,k,2)+vel_f(i,j,k,2)+m_half*divtau_o(i,j,k,2)););
-                    });
-                }
-            }
-            else if (m_diff_type == DiffusionType::Explicit)
-            {
-                Array4<Real const> const& divtau_o = ld.divtau_o.const_array(mfi);
-                if (m_advect_momentum) {
-                    amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-                    {
-                        AMREX_D_TERM(vel(i,j,k,0) *= rho_old(i,j,k);,
-                                     vel(i,j,k,1) *= rho_old(i,j,k);,
-                                     vel(i,j,k,2) *= rho_old(i,j,k););
-                        AMREX_D_TERM(vel(i,j,k,0) += l_dt*(dvdt(i,j,k,0)+rho_nph(i,j,k)*vel_f(i,j,k,0)+divtau_o(i,j,k,0));,
-                                     vel(i,j,k,1) += l_dt*(dvdt(i,j,k,1)+rho_nph(i,j,k)*vel_f(i,j,k,1)+divtau_o(i,j,k,1));,
-                                     vel(i,j,k,2) += l_dt*(dvdt(i,j,k,2)+rho_nph(i,j,k)*vel_f(i,j,k,2)+divtau_o(i,j,k,2)););
-                        AMREX_D_TERM(vel(i,j,k,0) /= rho_new(i,j,k);,
-                                     vel(i,j,k,1) /= rho_new(i,j,k);,
-                                     vel(i,j,k,2) /= rho_new(i,j,k););
-                    });
-                } else {
-                    amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-                    {
-                        AMREX_D_TERM(vel(i,j,k,0) += l_dt*(dvdt(i,j,k,0)+vel_f(i,j,k,0)+divtau_o(i,j,k,0));,
-                                     vel(i,j,k,1) += l_dt*(dvdt(i,j,k,1)+vel_f(i,j,k,1)+divtau_o(i,j,k,1));,
-                                     vel(i,j,k,2) += l_dt*(dvdt(i,j,k,2)+vel_f(i,j,k,2)+divtau_o(i,j,k,2)););
-                    });
-                }
-            }
-        } // mfi
-    } // lev
-
-    // *************************************************************************************
-    // Solve diffusion equation for u* but using eta_old at old time
-    // *************************************************************************************
-    if (m_diff_type == DiffusionType::Crank_Nicolson || m_diff_type == DiffusionType::Implicit)
-    {
-        const int ng_diffusion = 1;
-        for (int lev = 0; lev <= finest_level; ++lev) {
-            fillphysbc_velocity(lev, new_time, m_leveldata[lev]->velocity, ng_diffusion);
-            fillphysbc_density (lev, new_time, m_leveldata[lev]->density , ng_diffusion);
-        }
-
-        Real dt_diff = (m_diff_type == DiffusionType::Implicit) ? m_dt : m_half*m_dt;
-        diffuse_velocity(get_velocity_new(), get_density_new(), GetVecOfConstPtrs(vel_eta), dt_diff);
-    }
+    update_velocity(StepType::Predictor, vel_eta, vel_forces);
 
     // **********************************************************************************************
     //
     // Project velocity field, update pressure
     //
     // **********************************************************************************************
-    ApplyProjection(GetVecOfConstPtrs(density_nph),new_time,m_dt,incremental_projection);
+    ApplyProjection(get_density_nph_const(),new_time,m_dt,incremental_projection);
 
 #ifdef INCFLO_USE_PARTICLES
     // **************************************************************************************
